Move metabox component to the classic-editor folder
